fix : 拖拽时，网格线偶尔消失

+++ photoedit/src/main/java/com/mcxtzhang/photoedit/widget/CropDragView.java
@@ -181,12 +181,12 @@ public class CropDragView extends View {
         canvas.drawRect(mCropRect, mCropLinePaint);
 
         // 4 lines
-        float horizontalStep = mCropWidth * 1.0f / 3;
-        for (float x = mStartX + horizontalStep; x <= mStartX + horizontalStep * 2; x = x + horizontalStep) {
+        int horizontalStep = (int) (mCropWidth * 1.0f / 3);
+        for (int x = mStartX + horizontalStep; x <= mStartX + horizontalStep * 2; x = x + horizontalStep) {
             canvas.drawLine(x, mStartY, x, mStartY + mCropHeight, mCropLinePaint);
         }
-        float verticalStep = mCropHeight * 1.0f / 3;
-        for (float y = mStartY + verticalStep; y <= mStartY + verticalStep * 2; y = y + verticalStep) {
+        int verticalStep = (int) (mCropHeight * 1.0f / 3);
+        for (int y = mStartY + verticalStep; y <= mStartY + verticalStep * 2; y = y + verticalStep) {
             canvas.drawLine(mStartX, y, mStartX + mCropWidth, y, mCropLinePaint);
         }
